tests/requirements.txt
pytest
pytest-testinfra
paramiko
